moved down-sensor message to aqi gauge instead of graph
