+++ src/components/AqiGraph/Util.tsx
@@ -27,7 +27,6 @@ interface GraphElement {
  */
 interface GraphProps {
   sensorDocId: string;
-  isValid: boolean;
 }
 
 /**
